+++ src/drivers/rtmouse.h
@@ -285,5 +285,7 @@ int rpi_gpio_function_set(int pin, uint32_t func);
 void rpi_gpio_set32(uint32_t mask, uint32_t val);
 void rpi_gpio_clear32(uint32_t mask, uint32_t val);
 void rpi_pwm_write32(uint32_t offset, uint32_t val);
+int gpio_map(void);
+int gpio_unmap(void);
 
 #endif // RTMOUSE_H
