+++ src/drivers/rtmouse_main.c
@@ -88,136 +88,20 @@ struct class *class_dev[ID_DEV_SIZE] = {
     [ID_DEV_MOTORRAWR] = NULL, [ID_DEV_MOTORRAWL] = NULL,
     [ID_DEV_MOTOREN] = NULL,   [ID_DEV_MOTOR] = NULL};
 
-static volatile void __iomem *clk_base;
-
 // used in rtmouse_i2c.c
 struct cdev *cdev_array = NULL;
 volatile int cdev_index = 0;
 
-// used in rtmouse_dev_fops.c
+// used in rtmouse_dev.c
 volatile void __iomem *pwm_base;
 volatile uint32_t *gpio_base;
 struct mutex lock;
 
 /* --- Static variables --- */
-// used in rtmouse_dev_fops.c
+// used in rtmouse_dev.c and rtmouse_spi.c
 #if LINUX_VERSION_CODE >= KERNEL_VERSION(5, 16, 0)
 struct device *mcp320x_dev;
 #endif
-
-/*
- * set function
- * called by buzzer_init(), set_motor_l_freq(), set_motor_r_freq() and
- * buzzer_write()
- */
-int rpi_gpio_function_set(int pin, uint32_t func)
-{
-	int index = RPI_GPFSEL0_INDEX + pin / 10;
-	uint32_t mask = ~(0x7 << ((pin % 10) * 3));
-
-	gpio_base[index] =
-	    (gpio_base[index] & mask) | ((func & 0x7) << ((pin % 10) * 3));
-
-	return 1;
-}
-
-/*
- * set mask and value
- * called by sensor_read(), set_motor_l_freq(), set_motor_r_freq(), led_put()
- * and motoren_write()
- */
-void rpi_gpio_set32(uint32_t mask, uint32_t val)
-{
-	gpio_base[RPI_GPSET0_INDEX] = val & mask;
-}
-
-/*
- * clear mask and value
- * called by sensor_read(), set_motor_l_freq(), set_motor_r_freq(), led_del()
- * and motoren_write()
- */
-void rpi_gpio_clear32(uint32_t mask, uint32_t val)
-{
-	gpio_base[RPI_GPCLR0_INDEX] = val & mask;
-}
-
-/*
- * pwm set function
- * called by buzzer_init(), set_motor_l_freq(), set_motor_r_freq()
- * and buzzer_write()
- */
-void rpi_pwm_write32(uint32_t offset, uint32_t val)
-{
-	iowrite32(val, pwm_base + offset);
-}
-
-/* --- GPIO mapping for Device Open/Close --- */
-/*
- * Get gpio addresses and set them to global variables.
- *   - gpio_base
- *   - pwm_base
- *   - clk_base
- *   - clk_status
- */
-static int gpio_map(void)
-{
-	static int clk_status = 1;
-
-#if LINUX_VERSION_CODE <= KERNEL_VERSION(5, 0, 0)
-	if (gpio_base == NULL) {
-		gpio_base = ioremap_nocache(RPI_GPIO_BASE, RPI_GPIO_SIZE);
-	}
-
-	if (pwm_base == NULL) {
-		pwm_base = ioremap_nocache(RPI_PWM_BASE, RPI_PWM_SIZE);
-	}
-
-	if (clk_base == NULL) {
-		clk_base = ioremap_nocache(RPI_CLK_BASE, RPI_CLK_SIZE);
-	}
-#else
-	if (gpio_base == NULL) {
-		gpio_base = ioremap(RPI_GPIO_BASE, RPI_GPIO_SIZE);
-	}
-
-	if (pwm_base == NULL) {
-		pwm_base = ioremap(RPI_PWM_BASE, RPI_PWM_SIZE);
-	}
-
-	if (clk_base == NULL) {
-		clk_base = ioremap(RPI_CLK_BASE, RPI_CLK_SIZE);
-	}
-#endif
-
-	/* kill */
-	if (clk_status == 1) {
-		iowrite32(0x5a000000 | (1 << 5), clk_base + CLK_PWM_INDEX);
-		udelay(1000);
-
-		/* clk set */
-		iowrite32(0x5a000000 | (2 << 12), clk_base + CLK_PWMDIV_INDEX);
-		iowrite32(0x5a000011, clk_base + CLK_PWM_INDEX);
-
-		udelay(1000); /* wait for 1msec */
-
-		clk_status = 0;
-	}
-
-	return 0;
-}
-
-/* Unmap GPIO addresses */
-static int gpio_unmap(void)
-{
-	iounmap(gpio_base);
-	iounmap(pwm_base);
-	iounmap(clk_base);
-
-	gpio_base = NULL;
-	pwm_base = NULL;
-	clk_base = NULL;
-	return 0;
-}
 
 /*
  * dev_init_module - register driver module
